refactor(idp): idp login button component name
